requirements.txt
altair
numpy
pandas
pydeck
pydub
streamlit
streamlit_webrtc
st-paywall
openai
aiortc
twilio
streamlit_modal
audio_recorder_streamlit
gtts
elevenlabs
google-auth
google-cloud-firestore
httpx-oauth